move default theme implementation into its own package
