+++ packages/framework/theming-ramp/src/index.ts
@@ -1,4 +1,3 @@
-export * from './Color';
 export * from './SettingsWorker';
 export * from './Theme';
 export * from '@fluentui-react-native/theme-types';
