+++ packages/theming/default-theme/src/index.ts
@@ -1,0 +1,1 @@
+export { defaultFluentTheme } from './BaselinePlatformDefaults';
